fow working

+++ Source/HslmRTS/Public/RTSAgentComponent.h
@@ -39,17 +39,17 @@ public:
 	void SetVisionRadius(int32 NewVisionRadius);
 	UFUNCTION(BlueprintCallable, Category = "Hslm|RTSAgentComponent")
 	void SetHeightLevel(int32 NewHeightLevel);
-	
+
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "RTSAgentComponent")
 	bool bAgentEnabled;
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "RTSAgentComponent")
-	int32 VisionRadius;
+	float VisionRadius;
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "RTSAgentComponent")
 	int32 HeightLevel;
 
 
 	TArray<int32>  CacheVisionTiles;
 	FIntVector CacheAgentTile;
-	
+
 	// FogOfWar END
 };
